Refactor some integration tests
* Reduce block expiries
* Move wumbo test to channel spec
